+++ CHANGES.rst
@@ -8,6 +8,7 @@ in progress
 - Caching: Set cache filename to appname. Thanks, @gtirloni.
 - Update to pandas 2.0 and requests 2.26
 - CI: Update to Grafana 8.5.21 and 9.4.3
+- Grafana 9.3: Work around delete folder operation returning empty body
 
 2023-07-30 0.15.2
 =================